duplicate code fix
